Make JSTP successfully install and work on Windows
JSTP failed to install on Windows if build had been unsuccessful
because of `cat` command invokation. However, even with that
fixed, or when the native addon builds successfully, it fails
to work on Windows when parsing Unicode strings. Since the only
use case for JSTP on Windows is development, where performance
doesn't matter as much as it does on a server, there's no point
in supporting the native addon for Windows and the solution is
to only use it on Unix systems and fall back to pure JavaScript
implementation on Windows.