Rename public Connection methods
Certain connection methods that send packets are now renamed to be readable
verbs or actions instead of corresponding JSTP packet names. The reason for
renaming the inspect method is that its name caused bugs when passing a
connection object into util.inspect (see
https://nodejs.org/dist/latest-v7.x/docs/api/util.html#util_custom_inspection_functions_on_objects).
Other methods are renamed for the sake of consistency.