doc: fix extra spaces inside links in README.md
Don't let the HTML parser convert the newlines into spaces that end up
being inside the links and getting highlighted by, basically, removing
those newlines. The code formatting is still there though, and the lines
are kept under 80 characters, it's just the style of the HTML code that
was changed.

Fixes: https://github.com/metarhia/jstp/issues/204
PR-URL: https://github.com/metarhia/jstp/pull/205
Reviewed-By: Denys Otrishko <shishugi@gmail.com>
Reviewed-By: Timur Shemsedinov <timur.shemsedinov@gmail.com>