test: enable tests for js parser implementation
Serde tests were only launched using the default parser implementation
(which should be the native one while developing locally, and is always
the native one on CI since it fails the build if it doesn't succeed
building the addon).  This commit makes the tests run with both
implementations, the default and the fallback JavaScript-only one.

PR-URL: https://github.com/metarhia/jstp/pull/258
Reviewed-By: Dmytro Nechai <nechaido@gmail.com>
Reviewed-By: Mykola Bilochub <nbelochub@gmail.com>