Version 2.0.0
Next major release of the package marks the addition of built-in
reconnection functionality, using UMD for browser bundle, support for
asynchronous SessionStorageProviders, support for custom logging on
client-side and support for custom authentication errors. CLI was also
updated to support some of the features above. Reconnection is enabled
by default and uses the binary exponential backoff algorithm; this is
considered a breaking change.

The other breaking change is decoupling of the parser and serializer
into the separate package: https://github.com/metarhia/mdsf. This means
that methods `parse()`, `stringify()` and `parseNetworkMessages()` are
no longer available as part of the jstp package.

Starting with this release `metarhia-jstp` is being moved into the
@metarhia organization on npm, the new name of the package is
`@metarhia/jstp`.

Notable changes:

 * lib: add support for async SessionStorageProvider (Mykola Bilochub)
 * lib: realize reconnection by transport replacement (Mykola Bilochub)
 * server: enable returning errors from authPolicy (Igor Gorodetskyy)
 * lib: add utility for call messages resending (Mykola Bilochub)
 * connection: enable custom logging on client (Mykola Bilochub)
 * build: use UMD for browser bundle (Alexey Orlenko)
 * lib,build: use Web Crypto API in browser (Mykola Bilochub)
 * serde: remove serde implementation and use mdsf (Mykola Bilochub)
 * cli: use new jstp features (Dmytro Nechai)