ws: use transport similar to the socket ones
Previously every JSTP message was sent in a separate WebSocket frame.
This commit changes WebSocket usage so that JSTP message is not limited
to a single WebSocket frame. This is accomplished by making the
WebSocket JSTP transport behave in a way similar to the other available
transports, which is after every message, there is a separator. This
change leads to a WebSocket frame being able to hold multiple JSTP
messages at once or a part of a big JSTP message, that cannot be
transmitted in a single frame. This is a breaking change since it
completely changes the way the messages are transmitted over WebSockets.